requirements.txt
flask
uvicorn
aiohttp
hypercorn
pynacl
flask[async]